ast: fix bug with non-deterministic directive error reporting
This fixes a bug with the error reporting for directives. Given that
errors are sorted by their origin `AstNodeId`s in the semantic stage,
this meant that directives would be reported in a non-deterministic
since they were not assigned an `AstNodeId` in the parser stage.

This patch wraps each directive name in a `AstNode` to automatically
assign it one, and this fixes the problem.