link: implement building the linker instance
This patch implements the general logic for building a linker
from the specified target options, user options and platform
specifics. The linker is built up, the arguments are added and
then the execution of the linker occurs.

There are still a few specifics that need to be polished off for
other operating systems rather than macOS targets, but it should
work in principle for all platforms!

This means that people can now use `hash build ...` to build
an executable from the command line, which is pretty cool!