ir: improve caching for data type lowering
It is often that data type lowering doesn't register a cache hit
when it should. This is because it uses the `DataDefId` and the `ArgsId`
as the cache key. However, when the arguments are empty, a new `ArgsId`
is always used which causes a cache miss.

The new mechanism will ignore the `ArgsId` if the arguments are empty and
just use the `MonoData` to lookup the data type in the cache. This
should avoid most redundant re-calculations of data types (especially
for the primitives).