source: introduce line range map data structure
A new data structure is introduced to represent a mapping from a range of
bytes to lines in a source file. This is useful for converting `Span`s
and `SourceLocation`s into exact line and column numbers by a simple
lookup. This is intended to be used by the compiler when generating
debug information during code generation, and when computing exact lines for
errors, and when figuring out line arrangements when formatting the code.