target: avoid awkward API for querying sizes/max/mins when handling bigints
Change `IntTy::size()` `IntTy::min()` and `IntTy::max()` functions to
avoid returning an `Option<..>`. It is now assumed that querying this
information on `ubig` and `ibig` types is an invalid operation, and hence
will panic.

Additionally, added `numeric_min()` and `numeric_max()` to compute the
min and maximum values in the form of a `u128`. This is useful for exhaustiveness
and other parts of the compiler where the min/max is needed in an encoded
form.