tir: introduce `HasNodeId` and re-implement directive handling
`HasNodeId` is a helper trait which specifies that a certain type
has an associated node-id on it, and hence `node_id()` can be
called to retrieve it. This is useful for things like `TermId`,
`FnDefId`, `Pat`, etc.

With this change, handling of `#run`, `#dump_tir`, `#pure`,
`#entry_point` have been re-implemented.

Additionally, lots of code cleanup around other places because
of this change.