lower: implement applying `repr` and add `discriminant` option
This patch reworks the way `repr` attributes are checked by
introducing a `hash-attrs` specific `ReprAttr` representation
which is parsed from a `Attr`. During parsing, various semantic
checks are perfomed, and can be emitted to the `AttrChecker`.

Additionally, this logic is re-used in `hash-lower` to actually
apply the attribute onto `Adt`s that specify a representation
attribute.

Furthemore, since `repr` supports `u8`, `u16`, ...  as well, this
meant that `discriminant` option on `AdtRepresentation` is now
also supported.