Fix crazy issue with stores in multithreaded contexts
Some store functions used to first read the current store length
using a read lock, and then proceed to write using a write lock. This
allowed other threads to also read the same value, and then proceed to
overwrite the other's value. This led to rare bugs in some tests.
This commit fixes that.