parser: re-work `SpanMap` to avoid high contention
Previously, when allocating a new node through `AstNode::new()` and
`AstNodes::new()`, we would make a call to `SPAN_MAP` in order
to store the `Span` of the node and receive a `AstNodeId`.
Whilst in practise this is fine, it means that we are constantly
contending on the `SPAN_MAP` lock, which isn't ideal, especially
with many `parser` workers. In fact, it often led to far poorer
performance with more threads than fewer threads.

To fix this issue, we introduce a `LocalSpanMap` which is used
by a `parser` thread to allocate `Span`s and `AstNodeId`s for
new nodes at the local level. Once the current parser job is
done, the `LocalSpanMap` is consumed and merged into the
global `SpanMap`. This means that writes to the global `SpanMap`
only occur in between jobs, and perform big bulk writes, instead
of many small writes.

Overall, this is a massive performance boost (especially with
higher number of threads). We're looking at a 3.5x performance
improvement with 4+ parser threads.