parser: avoid using a stack for `AstGenFrame`s
We can avoid using a stack for `AstGenFrame`s since we only
care about the current one and the parent one. This means
that we can change functions `new_frame` to simply "swap"
the current and old frames instead of pushing and popping
them.

This also means that functions in the ast-gen don't have to
manually `consume_frame()` anymore.