parser: Integrated new lexer into the parser
This commit switches over the sources of the `v2` lexer over the original
lexer source and hooks it up with the rest of the parsing pipeline.

This commit uses the new `v2` experimental lexer for the parser which
now accepts the "flat" version of the token stream. Since much work was
done before this commit to abstract away dealing with the token trees.
The migration was reasonably simple (with some minor span calculation
adjustments).