core: fix `NodeMap::add_module` out-of-order insertions
This fixes issue #980 where given that multiple imports are being
simultaneously parsed, the insertion order of the modules in `NodeMap`
is not correctly preserved when using a `Vec`. This comes from the
fact that the parser generates the corresponding `ModuleId`. Hence,
we respect the order of the insertion by using a map between the
generated `ModuleId` and the parsed `Module`s.