tests: setup `CompilerLogger` per test case
This is needed in to support the new policy of using `log::info!`
for logging and outputting useful compiler information to the user.
During testing however, we don't capture the logger and don't initialise
it with our custom logging interface. This commit changes that.

Also, due to some of the format changes in information output, some tests
are updated to reflect the new output.