messaging: use `CompilerOutputMessage` for `log::info!`s
This commit makes it that when `log::info!` displays data, that it
always goes through the `CompilerOutputMessage` format. This is so
that the outside message consumers always know the schema and hence
the format of the messages.

Also, added `@@Messaging` to places where they should emit `CompilerOutputMessages`s
when the format is decided.