lower: correctly handle string comparisons
Before, in some cases, we weren't correctly generating code
for string comparisons that were occurring in a direct `TirBinOp::BinOp`
case. This was happening since the calls were being lowered directly
as a `BinOp` and hence a simple comparison between the addresses was
occurring.

This patch now correctly invokes `LangItems::str_eq` when this case
occurs.