reporting: Improve `note_on_span*` functionality with `SpanGuard`
Allow for `note_on_span*` and friends to be used in a more ergonomic way
by allowing the caller to provide a `SpanGuard`. The guard is intended to
filter out note calls that are not relevant to the current span, i.e. if
the developer only wants to print on the entry point module, they can use
`SpanGuard::EntryPoint` to indicate that.