parser: consider physically separated chunks as separate singular expressions
This commit changes the parser's handling of singular expressions,
i.e. ones that don't have any operators between them. Previously, the
parser would eagerly continue parsing a singular expression until it
matched the legal structure of an expression. Whilst this was legal,
and handled accordingly at later phases of compilation, it certainly
created odd (and unexpected) behaviour in the parser.

Now, we consider singular expressions as a collection of tokens that
must be "physically" connected. For example, before hand we consider:

```rs
foo
()

// Or

foo ()
```

as a call to the function `foo`, however, now we consider them as
separate expressions. Now, calling foo can only be done with:
```rs
foo()
```

In addition to that, we also fix some newly "detected" issues in
existing tests.