Unify all crate imports under workspace `Cargo.toml`
Instead of defining dependencies all in their own `Cargo.toml` files (
including selecting features and specifying versions), we instead
unify all dependencies under the workspace `Cargo.toml` file. This
allows us to have a single source of truth for all dependencies, and
also allows us to use the same version of a dependency across all
crates. This is especially useful for crates that are used in multiple
places, as it ensures that we are always using the same version of the
dependency.