parser: standardise naming conventions for parser closures
Always use `g` for the name of the generator passed into parsing
closures, i.e.
```rust
self.in_tree(..., |g| {
//                 ^ Always use `g` here
    ...
})?;
```

This renaming is also done to align with the fact that `gen` is becoming
a reserved keyword in Rust edition 2024. This change is made to prepare
for upgrading to the 2024 edition.